feat: move gcgraph to detail
