+++ modules/imgproc/src/grabcut.cpp
@@ -40,7 +40,7 @@
 //M*/
 
 #include "precomp.hpp"
-#include "gcgraph.hpp"
+#include "opencv2/imgproc/detail/gcgraph.hpp"
 #include <limits>
 
 using namespace cv;
